revert changes in examples  this_that_svc, puppeter
