+++ examples/this-that-svc/wasm/Cargo.toml
@@ -10,12 +10,5 @@ this-that-svc-app = { path = "../app" }
 
 [build-dependencies]
 gwasm-builder.workspace = true
-sails-client-gen.workspace = true
 sails-idl-gen.workspace = true
 this-that-svc-app = { path = "../app" }
-
-[dev-dependencies]
-gtest.workspace = true
-sails-rtl.workspace = true
-parity-scale-codec = { workspace = true, features = ["derive"] }
-tokio = { workspace = true, features = ["rt", "macros"] }
