+++ examples/this-that-svc/wasm/build.rs
@@ -1,6 +1,6 @@
-use sails_idl_gen::program;
+use sails_idl_gen::service;
 use std::{env, fs::File, path::PathBuf};
-use this_that_svc_app::Program;
+use this_that_svc_app::MyService;
 
 fn main() {
     gwasm_builder::build();
@@ -9,11 +9,7 @@ fn main() {
 
     let idl_file_path = manifest_dir_path.join("this-that-svc.idl");
 
-    let idl_file = File::create(idl_file_path.clone()).unwrap();
+    let idl_file = File::create(idl_file_path).unwrap();
 
-    program::generate_idl::<Program>(idl_file).unwrap();
-
-    let out_dir_path = PathBuf::from(env::var("OUT_DIR").unwrap());
-    let client_rs_file_path = out_dir_path.join("this_that_svc_client.rs");
-    sails_client_gen::generate_client_from_idl(idl_file_path, client_rs_file_path).unwrap();
+    service::generate_idl::<MyService>(idl_file).unwrap();
 }
