--- examples/this-that-svc/wasm/tests/this_that_svc_client/mod.rs
@@ -1,2 +1,0 @@
-#![allow(dead_code)]
-include!(concat!(env!("OUT_DIR"), "/this_that_svc_client.rs"));
