+++ examples/this-that-svc/wasm/this-that-svc.idl
@@ -9,7 +9,7 @@ type ManyVariants = enum {
   Two: u32,
   Three: opt u256,
   Four: struct { a: u32, b: opt u16 },
-  Five: struct { str, h256, nat64, nat256 },
+  Five: struct { str, h256 },
   Six: struct { u32 },
 };
 
@@ -17,11 +17,7 @@ type TupleStruct = struct {
   bool,
 };
 
-constructor {
-  New : ();
-};
-
-service ThisThatSvc {
+service {
   DoThat : (param: DoThatParam) -> result (struct { str, u32 }, struct { str });
   DoThis : (p1: u32, p2: str, p3: struct { opt str, u8 }, p4: TupleStruct) -> struct { str, u32 };
   Noop : () -> null;
